requirements.txt
aiogram==3.17.0
yt-dlp==2025.1.26
python-dotenv==1.0.1
aiofiles==24.1.0